Move adj matrix to representations
